requirements.txt
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.2
pytz==2022.6
scikit-learn==1.2.1


requests~=2.28.2

yfinance~=0.2.18
django-scheduler~=0.10.1
keras~=2.12.0
tensorflow~=2.12.0
pymongo~=4.3.3